refactor: Update Snort service configuration to include log directory parameter
The Snort service configuration file has been updated to include the `-l` parameter, which specifies the log directory for storing Snort logs. This change ensures that Snort logs are stored in the specified directory.

Refactor the `install-snort3.sh` script to modify the `ExecStart` line, adding the `-l` parameter with the value of `$LOG_DIR`. This change allows Snort to write logs to the correct directory.

This commit addresses the issue of Snort logs not being stored in the desired location and improves the overall functionality of the Snort service.